Fix organize files (#8)
* install node packages

* remove firebase and files from project

* modify infocolumns component name

* remove firebase code from header component

* modify infocolumns implementation name and remove unused component

* modify style files

* remove unused imports

* import styles

* modify styles import route
