+++ package.json
@@ -7,8 +7,6 @@
         "@testing-library/react": "^13.3.0",
         "@testing-library/user-event": "^13.5.0",
         "bootstrap": "^5.1.3",
-        "firebase": "^9.8.3",
-        "firebase-tools": "^11.0.1",
         "react": "^18.1.0",
         "react-bootstrap": "^2.4.0",
         "react-dom": "^18.1.0",
