--- src/components/info-columns/InfoColumns.jsx
@@ -1,22 +1,0 @@
-// @vendors
-import React from 'react'
-import { Container, Row } from 'react-bootstrap'
-
-// @components
-import { InfoGrid } from '../ui/InfoGrid'
-
-// @static
-import data from '../../data/dataColumnsInfo.json'
-
-export const InfoColumns = () => {
-    console.log(data.gridData)
-    return (
-        <Container className="section ">
-            <Row>
-                {data.gridData.map((item, index) => (
-                    <InfoGrid icon={item.url} key={index} text={item.text} />
-                ))}
-            </Row>
-        </Container>
-    )
-}
