--- src/components/info-columns/index.js
@@ -1,1 +1,0 @@
-export * from './InfoColumns'
