+++ src/components/infoColumns/infoColumns.sass
@@ -1,0 +1,12 @@
+.section
+    background: #e5e5e5 !important
+    margin: 0
+    max-width: 100vw
+    padding: 5.75rem
+    width: 100%
+
+    &:first-child
+        margin-left: 0
+
+    &:last-child
+        margin-right: 0
