+++ src/components/ui/InfoGrid/grid.sass
@@ -1,0 +1,29 @@
+.box
+    align-items: center
+    background-color: #c1d7d3
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25)
+    border-radius: 25px
+    display: flex
+    flex-direction: column
+    font-size: 24px
+    justify-content: flex-start
+    z-index: 2
+
+    .card
+        background-color: #c1d7d3
+        background-clip: none
+        border: none
+        border-radius: none
+
+        &-header
+            background-color: transparent
+            border-bottom: none
+
+    .grid-icon
+        /* bottom: 295px */
+        filter: drop-shadow(4px 4px 4px rgba(0, 0, 0, 0.25))
+        /* left: 205px */
+        /* position: absolute */
+        margin-top: -90px
+        width: 170px
+        z-index: 2
