+++ src/index.js
@@ -1,11 +1,11 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import "./styles/sass/index.scss";
-import App from "./App";
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import './styles/index.scss'
+import App from './App'
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-);
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>
+)
