+++ src/styles/components/_header.scss
@@ -5,7 +5,6 @@
     font-size: 30px;
     line-height: 37px;
     margin-bottom: 43px;
-    text-justify: center;
 }
 
 .wave-container {
