+++ src/styles/index.scss
@@ -1,8 +1,6 @@
 @import '~bootstrap/scss/bootstrap';
 @import './pages/App';
 @import './components/header';
-@import './components/infoColums';
-@import './components/grid';
 
 * {
     margin: 0px;
